requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
circuitbreaker==1.4.0
click==8.1.3
cryptography==3.4.7
Flask==2.1.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oci==2.75.0
psycopg2-binary==2.9.6
pycparser==2.21
pyOpenSSL==19.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.7.0
urllib3==1.26.10
waitress==2.1.2
cryptocode~=0.1
flask-restx>=0.0.0